Implemented methods for cluster
Also replaced any occurrences of 4 spaces in game.js with tabs, moved
worker code for cluster to a separate file.
The methods need to be refined to properly be used, including changing
timeouts to intervals and moving methods for pathfinding to worker.js or
a different file before being implemented.