Added basic lag compensation
The lag compensation could be a lot better done, like moving the object
along the path it is supposed to instead of just moving it towards its
target. But, it does make the game run smoothly.

Side note: The game is almost at POC level thanks to this addition! It
just needs some class balancing, and have buildings actually do things.